+++ CHANGELOG.md
@@ -28,6 +28,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Detailed analysis for new constraints in GUI.
 - Metadata information handling for each instance processed.
 - Visualization graphs for single and multiple instances (accumulated) over the number of instances solved.
+- Proper Visualiation graphs for solution statistics, efficiency metrics, constraint analysis.
+- Cache clearing for accumulated solved instances for visualisation.
+- New plots tab for the GUI to directly visualize the each of the saved files for the user.
 
 ### Changed
 - Updated the GUI `views.py` and `controllers.py` to fit in to our new solver solution (`ortools-fast.py` but now in a dynamic and versatile way).